requirements.txt
Flask==2.3.2
langchain==0.0.231
openai==0.27.8
google-search-results==2.4.2
PyMySQL==1.1.0
python-dotenv==1.0.0